rename tt -> mbe, part 2
